requirements.txt
asgiref==3.6.0
Django==4.2
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2